Fix issue with specs failing for warden
Previosuly we were loading in the devise controller helpers in all
spec files, but this caused some internal spec files in devise to
error out.

This fix changes the devise helpers to only load for controllers,
and updates the application helper spec to stub out current_user
instead of simulating a log-in with the sign_in helper.