Replace Armoire with Dotenv
Previously we were using Armoire for handling the configs in the app,
but we ran into issues when getting TravisCI to read the database.yml.

This PR updates our app to use Dotenv so we can store the username and
password for the database in ENV variables instead, which we can then
set in TravisCI's settings.